zsh: expand bmc_editor to handle detachable editors
Some editors, like graphical Vim variants, can detach from the terminal.
This is not useful for the EDITOR or VISUAL environment variables, but
it is useful for general purposes usage, and being able to pick such an
editor is valuable because it can then be suitably aliased.

Let's expand bmc_editor to handle a --detach argument, which asks for an
editor which can detach from the terminal.  We handle the case where we
have the same editor with no argument and an argument by checking to see
whether the editor has an arguments and factoring that into our decision
using our lookup table.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>