zsh: alias gvi to a detachable editor
We'd like to provide for an alias to use the best detachable editor
possible.  Let's ask bmc_editor to provide such an option and alias it
to "gvi".

If there isn't one, let's not create the alias.  To do that, we want to
choose whether we want to fall back to vi (in this case, we do not) and
avoid creating it in that case.  In order to make this easier, add
--fallback and --no-fallback options to bmc_editor to decide whether a
fallback is desired.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>