zshrc: fix terminal overrides
The current special handling for TERM using SSH and Docker
short-circuits the main code whose job it is to determine if the
terminal definition we're using is working.  As a result, if our TERM is
set to something new like tmux-direct and we SSH into another system,
the code automatically fixing the terminal definitions doesn't run due
to the special handling returning zero.

Instead, let's make the main overrides run and then specially handle the
case where we set up COLORTERM for SSH or Docker, since we won't know
until that point whether we have a suitable terminal definition.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>