Makefile: set permissions explicitly
Right now, we always clear the permission bits for other, but we don't
control group and user permissions.  Some programs, such as GnuPG, don't
like it when anyone other than the user can access the directory, and
installing the gpg.conf file can cause subsequent invocations to
complain.

In any event, it's not particularly useful to let people even in our
group have access to our dotfiles.  On Linux, we have private groups, so
nobody else would have access to them, but on macOS, that would give
every user on the system access to read our dotfiles, which wouldn't be
a good idea.  Let's explicitly set permissions only for the user.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>