playbooks: add an Ansible playbook to set up a Codespace
The default Codespaces environment has a way to deploy dotfiles, but it
doesn't consider that a user might have different dotfiles for home and
work.  If the default dotfile branch is for home, then it makes it very
difficult to automatically deploy the right set of dotfiles for work,
since the setup script in the default branch would have to have special
logic for both branches to determine the right environment.

Let's make it easy to set up a new Codespace by using Ansible and create
an Ansible playbook to do this automatically.  Ansible is well known to
be a robust tool for remote automation using nothing more than SSH, and
it's clearly up to the task.

Signed-off-by: brian m. carlson <bk2204@github.com>