zshrc: unset TMOUT
Some systems "helpfully" set a shell timeout to log users out after a
period of inactivity.  This causes things like long-running shell
sessions which must run overnight in tmux to die even if they contain
useful debugging information, and is not in fact very helpful at all.

Use `typeset` to mark this variable as not readonly and proceed to unset
it.  Note that setting it to zero would be equivalent here, and also
that this only affects interactive shells.  Thus, setting it in
`.zshenv` is not necessary because non-interactive uses like scripting
won't be affected.