vimrc: switch from rls to analyzer
The preferred LSP server for Rust is now Rust Analyzer.  It's faster and
more robust in the face of errors and it's now included with rustup.
Since Rust Analyzer is better in every way and RLS is deprecated and
scheduled to be removed, switch ALE to use Rust Analyzer instead of RLS.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>