codespace-setup: only run Git commands in Git directories
In some cases, we can end up with code under /workspaces that is not in
a Git repository.  In such a case, attempting to run a Git command isn't
going to work and is just going to lead to the failure of our play.  Fix
this by verifying that the directory looks like a Git repository before
trying to process it.

Note that there are some more accurate ways to do this check, but for
now this seems to work.  If we need a better option in the future, we
can always improve on things then.