codespace-setup: get credentials for git remote set-head
In a codespace, we have some credentials provisioned for us.  These are
provisioned using a script in `/etc/profile.d`.  This is a good way to
make these credentials available to all POSIX shells, but `/etc/profile`
and friends are only read for interactive shells.  There is no portable
way to specify shell setup for noninteractive POSIX shells, however.

This poses a problem for us when we're running our play, because Ansible
obviously doesn't spawn an interactive shell.  As a result, it lacks the
environment variables it needs, such as $GITHUB_TOKEN, and it can't
invoke `git remote set-head`.  Let's work around this by sourcing the
proper entry in `/etc/profile.d` so that we can use the credentials we
need.

Signed-off-by: brian m. carlson <bk2204@github.com>