dct-cs: allow looking up multiple facts at once
Right now, we get only the username from the remote server.  However,
we'll also want to get the user ID and the existence of the lawn binary
on the remote server as well.  When we do that, let's store the entire
data we've gotten in a variable and then parse it out as we need it.

This avoids having to duplicate the data or make multiple SSH requests.

Note that because the SSH protocol only sends a single string over the
wire which is interpreted by the shell, we use `git rev-parse
--sq-quote` to turn it into something that results in a properly quoted
string on the other side.