Xsession: set default Qt theme
If we're using MATE, let's set the default Qt theme to be one using GTK+
for rendering.  Note that the GTK+ 3 engine doesn't include styles, so
we need to use the GTK+ 2 setting.  If the appropriate package is not
installed, then automatic fallback happens to the default theme, so
everything will continue to work.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>