mutt/config/iface: set status colours to cyan and black
The yellow on blue can be hard to see in some colour schemes, such as
the Tomorrow Night scheme used by default for Alacritty.  Change the
colour scheme to cyan on black, which should be more distinct and easier
to see.

Note that Mutt does support 256 colors and we could choose those, but
only when the number of colours is exactly 256.  This means that with a
true-colour terminal, we can't use the 256-colour palette integers, so
we need to stick with the standard 16-colour palette.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>