zshrc: remove xmodmap
xmodmap is automatically invoked in our Xsession file at startup time,
so it should be set automatically.  However, we also currently invoke it
any time the shell is started, which on newer versions, causes messages
to be printed if any keys are held down.  This means that starting a
tmux session and then quickly pressing Alt-Tab can cause many terminals
to all print this messages, which is undesirable and ugly.

Since xmodmap should be invoked automatically on startup, let's avoid
invoking it in every shell to avoid this unwanted message.