zshenv: ignore inability to set core ulimit
On Cygwin, setting a core ulimit isn't possible, so we print an error
when starting the shell.  Not having this limit is fine, because if
there's no core files, then we don't have to do anything to disable
them.  Let's check to see if the value exists and if so, set it; if not,
we do nothing, silently.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>