zsh: make zprofile load zshenv
In our configuration, zshenv loads all of the environment variables we
intend to use generally, including PATH.  However, on some systems, the
system profile code overrides PATH, leaving us with a broken PATH that
doesn't include anything we've set.  Notably, this occurs on a Codespace
if PATH is set by the Codespace.

To solve this, have zprofile, which runs after the system profile code,
simply source zshenv, which will stuff the overridden PATH code at the
end instead.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>