zsh: append `--` to nvim-gtk commands
Neovim-GTK accepts two sets of options: options for itself, and after
`--`, options for Neovim.  Most programs that invoke editors assume that
one can jump to a line by writing `+` followed by the line number, but
that doesn't work with the current settings, since Neovim-GTK doesn't
know how to handle it and doesn't pass the data off to Neovim itself,
which does.

To help it out, let's include the `--` value after the rest of the
editor text to make it pass any remaining args to Neovim.  This should
make the `+` syntax as well as any other Vim-like arguments just work,
and in the extremely unlikely situation we actually want to customize
the Neovim-GTK arguments, we can just invoke `nvim-gtk` directly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>