vimrc: add a key combination to copy filename to register
It can often be useful to get the filename of the current buffer for
copying and pasting.  Add two key combinations that copy the current
buffer filename into a register, which could be one of the clipboard
registers.  `<Leader>zf` copies the relative path, while `<Leader>zF`
copies the absolute path.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>