zsh/bmc_palette: don't overlap true colour and 256 colours
Right now, we use the first two bytes of the hex value of the hostname
as an index into the 256-colour palette as well as for the red component
of a true colour value.  Let's avoid overlapping these values by using
the first six hex characters as the true colour value and the next two
hex characters as the 256-colour palette index.  That way, we can rely
on colours being truly independent.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>