zsh: improve colour handling
Right now, it's nearly impossible to specify that we'd like to use a
256-colour value instead of a true colour value.  Let's refactor the
colour handling slightly and instead allow the true colour handling to
accept the value skip, in which case, we'll emit a 256-colour value.

This requires in some cases that we ignore the fact that our terminal is
more capable.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>