misc/alacritty: use Source Code Pro
Now that Alacritty 0.12 has made its way to Debian experimental, let's
switch the font to Source Code Pro.  The version in Debian, which is
shipped with TeXLive, is broken in a way that causes Alacritty 0.11 to
print all of the text over itself at the top of the screen.  Alacritty
0.12 handles this gracefully by rendering the text as expected.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>