vimrc: disable u/U in visual mode
In visual mode, it's possible to type "u" to make a text lowercase, or
"U" to make it uppercase.  This contrasts with normal mode, where "gu"
or "gU" is required.

Unfortunately, on a standard QWERTY keyboard, "u" is right next to "y",
so in visual mode, it's very easy to make text lowercase instead of
yanking it into a register, which is typically not what was intended.

Let's make it harder to do the wrong thing by disabling the "u" and "U"
motions in visual mode so that typing mistakes for "y" don't lead to
unexpected behaviour.  Note that "gu" and "gU" work just fine, just as
in normal mode, so no functionality is lost.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>