zsh: adjust colour handling to use options
Right now, our colour handling for the terminal palette assumes that we
always have a 16-colour value, a 256-colour value, and a true colour
value.  We also use a helper function to decide which of these options
is the correct one and emit it.  This leads to complex, hard-to-read
code.

Instead, let's use zsh's option parser in `color_fg` and `color_bg` to
extract the values we _do_ have, and then emit the proper value based on
what was provided to us and how many colours we have.  This gets rid of
the need for `color_count_for`, since the logic is simple and easy to
inline.  This also allows us to prefer a 256-colour value instead of a
true colour one if we don't feel like providing the latter.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>