playbooks: deploy Lawn
Lawn is a tool to make working with remote systems easier, such as by
spwaning processes on the local machine.  In order to make it possible
to work with it in Codespaces, deploy it to the Codespace.

In addition, create a directory to forward sockets to and a directory to
mount remote file systems.

Signed-off-by: brian m. carlson <bk2204@github.com>