playbooks: add a script to set up repositories
The current approach of using `git remote rename` performs poorly with
many refs because for each reference that has to be renamed, rewriting
the packed-refs file is an O(N^2) operation.  We can do this more easily
with `git update-ref` if we have a new version of Git with transactions,
but the command is long and doesn't belong in an Ansible playbook.

To solve this problem, let's add a snippet that does all of our
repository setup and use that to fix our repositories the way we want.
Replace the multiple lines in the Ansible playbook with a single one
that calls this snippet.

Signed-off-by: brian m. carlson <bk2204@github.com>