vimrc: enable Sorbet out of current directory when safe
Sorbet provides type checking and an LSP for Ruby.  Normally, our LSPs
don't come from the root of our repository, so we don't need to worry
about malicious content.  However, for Sorbet, this is different, since
it's usually included as a program in the Gemfile.

If we're using an external project, one that we're never going to
contribute to, then don't start Sorbet from the repository root.  Only
start it if it's present and the repository is in our normal, trusted
repository directory.  This way, we won't accidentally execute arbitrary
code.

In such a case, enable Sorbet for fixing as well, since it often has
fixes that can be adopted.

Signed-off-by: brian m. carlson <bk2204@github.com>