zshrc: don't load environment from tmux if not in tmux
If we're not in tmux, attempting to load the environment from the shell
that started tmux is simply going to result in us overriding important
variables.  For example, if we started a tmux session, then detached and
logged out, loading the SSH socket variable from tmux results in us
using the old socket from the last session, which no longer exists.

To avoid this, only allow us to load the environment if `$TMUX` is set,
and otherwise, fail.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>