Implement a browser wrapper
The `BROWSER` environment variable is a colon-separated list of browser
names.  Some implementations pass these arguments to the shell, allowing
multiple arguments, and some do not.  When we're on a remote machine, we
want to invoke Lawn, so that we can run GUI items on a local machine,
which means we need multiple arguments.

In addition, on Ubuntu, the Firefox snap spews lots of GTK+ warnings to
standard error, and Mozilla doesn't seem to be interested in fixing this
promptly, nor do the GTK+ developers care that the terminal is being
polluted.

To resolve these concerns, implement a wrapper, `dct-browser`, that
simply invokes the desired browser, choosing from the first of the
ordered options.  If the Ubuntu Firefox snap is in use, then filter out
all of the GTK messages from standard error so our terminal stays nice
and tidy.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>