zshenv: prevent global compinit
On Ubuntu, `/etc/zsh/zshrc` tries to invoke `compinit`.  While
convenient, if we're running under `sudo -E`, this fails due to load our
shell completion because it has the wrong permissions.  _Our_ version of
this uses the `-u` flag, which works correctly, but the Ubuntu version
does not, and so it prints an ugly, interactive error message every time
it's loaded.

Silence this by not using the global compinit if we're on Ubuntu.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>