vimrc: accept #/ comments as well as # comments
It's often useful to include help and usage output in scripts.  In
languages using `#` as the comment character, this can be easily
accomplished by writing the help output in a specially formatted comment
line and simply extracting it from the script file at runtime.  This is
especially useful in shell scripts, where the task can be accomplished
easily with standard shell tooling.

Use `#/` as the form of this special comment, which is the form that
GitHub uses for this, and teach Vim to insert it automatically.  As the
documentation directs us, we specify the longer comment character before
the shorter version.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>