vimrc: check for Neovim before `gui_running`
In Neovim 0.7.2 and older, `has("gui_running")` is always false.
However, in at least 0.9.4 and newer, it can be true if a GUI is
enabled.  This is actually a good thing, but it poses some problems.

Our `SetGUIOptions` function sets the font, which for Neovim, looks like
a call to `Guifont`.  However, neovim-gtk hangs when `Guifont` is
invoked directly from the `.vimrc`, before the `UIEnter` event is
fired.  Thus, with the current ordering of things, we'll invoke that
command due to `has("gui_running")` being true, and thus we'll hang.

To solve this, let's first check for Neovim, and if we have it, then
simply set the options on, and only on, the `UIEnter` event.  If we're
not using Neovim, then we must be using Vim, and so calling the setup
code directly is okay.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>