Allow importing additional rules
Right now, all dotfiles are public.  However, there may be some
functionality which makes sense to include in the dotfiles but cannot be
distributed publicly under an open source license.  Examples are
general scripts which have no clear license or scripts which may reveal
internal work details.

Let's add an additional, optional target to include in the main
Makefile.  This makes it possible to add an overlay in a separate
repository which can contain rules to install those files even though
they're not included here.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>