bin: add a tool to process files through ERB
Right now, we have some dotfiles which might need to change on different
systems.  For example, the Super key usage in the Alacritty
configuration would need to change on macOS.

Let's add a tool to process files through ERB based on a YAML file.
This allows us to generate configuration files based on things like the
OS or a simple YAML file.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>