zsh: cache colours on initial setup
We call `bmc_palette` many times during the prompt and zsh process to
set up various colours.  However, we can do better by making
`bmc_palette` a setup function and calling it once, which will then
cache the values in a variable.  This means we compute each value once
and can reuse it cheaply.

Adjust the `bmc_palette` file to call the setup function and autoload it
so that we don't need to call it once the autoload happens.  Declare the
variable we use for caching globally so it can be used by all the
relevant functions.

Because we're no longer loading the hostname only on demand, remove the
conditional surrounding it.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>