zsh: avoid needlessly calling tr
POSIX specifies that a command substitution strips trailing newlines.
The only case in which our `tr -d` would have any effect is if we had
non-trailing newlines, which would be a bizarre output from the programs
we've invoked.  To save startup time and avoid needless computation,
let's just skip the `tr`.  In the unlikely event we do get some bizarre
output with an internal newline, we'll just go with it and compute the
colour accordingly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>