zshenv: avoid calling `has_locale` many times
`has_locale` calls Perl, which has a nonzero startup cost.  This, in
turn, is called by `preferred_locale`, which we call five times.  The
set of locales is not going to change in between our calls, and neither
is the locale we choose as most preferred.  To avoid needless work,
let's compute the value once, assign it to a variable, use the variable
where we need it, and then, to keep things tidy, unassign it before we
leave our file.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>