zshrc: minimize calling `silent`
`silent` is a very convenient wrapper function, but it's not free.
Specifically, when we set up terminal settings, we call `echotc` many
times, which can have a cost.  In addition, because `echotc` is a
built-in and requires redirection of standard error anyway in order to
not scream to the screen, let's just simplify and avoid the use of
`silent` here.  This speeds up our startup a bit.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>