zsh: avoid recomputing parent process info
Right now, we compute the parent process's executable name in every
precommand.  However, this is unnecessary.  The value only matters for
interactive shells, and it won't change without a new interactive shell
being spawned.  Since any new interactive shell would set up the prompt
itself, we can compute it once with the prompt and not have to reuse it.

Let's do that, and we'll also apply the optimization from the previous
commit where we use the shell to call `realpath` on `/proc/$PPID/exe`
when using Linux, rather than calling out to `ps`, which also works, but
is slower.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>