Makefile: add targets to template and build generated config files
Right now, we have no way to build a generated config file.  For
example, if we want to change an email address, it must be changed in
several places, which is inconvenient and error prone.

Instead, let's add an appropriate templating mechanism so we can build
and install generated files from an ERB template.  It remains optional
because we won't always have Ruby to perform these actions.

If `TEMPLATE` is set to 1, then templating is available; otherwise, it
is missing.  This allows various `rules.mk` files to adjust their
behaviour based on the availability of this functionality.

Label all of the files with the `.gen` suffix to make them easy to
ignore, and then ignore them so we don't check them in accidentally.
Also ignore the config.yaml file, which will be specific to the
particular machine and environment we're running on.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>