git: add a templated `~/.config/git/config` file
It's useful to have some per-environment settings in Git, such as
whether commit signing is enabled and which credential helper to use.
Read this data from the configuration file and generate a secondary
Git config file in the XDG-compatible location.

If the credential helper is specified as `os`, then pick the right one
for the OS that we're on.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>