tmux: set escape time to 100 ms
Neovim recommends an escape time that is 300 ms or less here,
whereas the default is 500 ms.  Let's set it to 100 ms, which is a tenth
of a second, which should be sufficient to determine if the user is
smashing escape or there's an escape sequence.  It's very unlikely that
a user will be capable of hitting Escape twice within 100 ms, which is
exactly what we want.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>