tmux: use /bin/sh as shell for overmind sessions
Our zshenv unconditionally overrides the PATH to include some default
values at the front, including the standard paths, such as /usr/bin.  If
we invoke overmind with a specific PATH value that includes, say, a
different `node` version at the front of the PATH, it invokes tmux with
the default settings, which invokes zsh, which sets the PATH back to
have /usr/bin before that special path.  As a result, we invoke the
wrong binary.

We don't really want to override the settings for invoking processes
just because overmind uses tmux, so let's create a special tmux
configuration file to use `/bin/sh` as the default shell.  That won't
override any of our PATH settings.  To make this effective, set
`OVERMIND_TMUX_CONFIG` in the environment so that we use our new tmux
config.

Signed-off-by: brian m. carlson <bk2204@github.com>