playbooks: unminimize Ubuntu codespaces
Ubuntu containers are minimized to not contain documentation, including
manual pages, on the theory that this saves space for systems which a
user will never log into.  Unfortunately, for a devcontainer, people do
in fact log into them, and it's quite irksome not have documentation.

Unminimize our image as one of the first things we do, so that we have
the documentation.  Note that it's important to do this early, since it
involves reinstalling lots of packages, and thus it's more efficient to
do it before installing additional software.

Signed-off-by: brian m. carlson <bk2204@github.com>