Makefile: allow customizing configuration file
Right now, we always check for `config.yaml` in the root of the
repository as our configuration file.  However, we'd like to be able to
add some tests in the future, and to test different configurations,
we'll need to use a custom configuration file.

Let's add a variable to allow us to customize the configuration file and
check for that file in all the normal situations.  In addition, add a
`print` target for debugging purposes.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>