git: make template work gracefully if no config values specified
We want to do something reasonable when templating if our configuration
file contains no actual values, because this means that we'll also be
resilient to changing values or incomplete configuration.  Let's use
Hash#fetch to provide a default in this case.

Now that we have this working, let's add some tests to make sure it
continues to work.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>