Add a snippet to automatically unsubscribe from emails
RFC 8058 specifies a one-click unsubscription option for emails, and
recent Gmail updates have basically made it mandatory for all major
companies and mailing lists.  Let's add a snippet that can accept a
compatible email on standard input and automatically unsubscribe.

Note that in our Perl snippet, we need to specifically look for the
`List-Unsubscribe` header, because we will also have
`List-Unsubscribe-Post` header which formail will include, and we want
to look case insensitively because mail headers are case insensitive.

We also need to specifically look for an HTTPS URL, because there may
also be `mailto` or other schemes as well, which we will not want to
use.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>