vimrc: invoke Sorbet with bundler
In some development environments, Sorbet does not work unless invoked
from bundler, returning an error like the following:

    lib/ruby/3.3.0/bundler/spec_set.rb:248:in `each': undefined method `name' for nil (NoMethodError)

          @specs.sort_by(&:name).each {|s| yield s }
            ^^^^^^^^

In such a case, the solution is usually to run it under bundler.
However, ALE will not accept multiple arguments, so add a shell script
that invokes it.  In addition, use the project's version of bundler,
since the default system bundler doesn't work correctly.

Signed-off-by: brian m. carlson <bk2204@github.com>