Don't use OS credential helper in a codespace
If we're in a codespace, we probably don't have the libsecret credential
helper present, and in any event, we couldn't access the relevant data
because it requires gnome-keyring or equivalent to be running, which
isn't the case in a codespace.

Let's default to no particular credential helper and then we won't set
the option.  When Lawn learns how to provide support for a credential
helper in the future, we can substitute that.