mutt: add "E" to unsubscribe from a message
Google and Yahoo have recently requires one-click unsubscribe for most
senders, which means that as a practical matter, almost all emails
related to marketing or mailing lists support this.  Add a key, E, to
invoke a simple script that makes an appropriate unsubscribe call based
on the email piped on standard input.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>