Neuter Azure CLI surveys and telemetry
The Azure CLI surveys and telemetry should already have been fixed by
the appropriate configuration file that we've installed.  However, we
want to be absolutely sure that any bug in the tool that might cause it
to accidentally send surveys or telemetry will cause it to fail.

To do so, symlink these files to `/dev/null`, which will hopefully make
the tool puke or get bad data if it tries to invoke the forbidden
functionality.

Signed-off-by: brian m. carlson <bk2204@github.com>