alacritty: add a TOML version of config file
Newer versions of alacritty have switched to TOML.  Let's add a version
of the config file in TOML format so that Alacritty doesn't complain.
Organize the file similarly to the YAML version.

Leave the YAML version around until Debian does a stable release with a
version that uses TOML instead of YAML.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>