Add a program to play sounds
Playing a sound, especially a system beep, in a cross-platform way is
difficult.  We want this functionality for Alacritty, which has no
method of playing a beep built-in.  Add a command that finds the proper
sound file if necessary and invokes a suitable program to play it.

Use this functionality in Alacritty.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>