Add a tool to process mtree files
Most of the BSDs contain a utility called `mtree`, which can produce a
manifest of files on the file system, their ownership and permissions,
and optionally other metadata, including their cryptographic hash.  This
utility is also part of Debian and other distros, and similar
functionality is included in libarchive.

Our current installation approach for dotfiles lacks some features.
First, it is completely ignorant of permissions, and second, it requires
symlinks to be located in the repository.  Finally, it cannot handles
files with spaces or shell characters because installation occurs with
`make`.

To fix this, let's add a utility that can process mtree files from in
the repository and install data into the proper location based on such a
manifest.  We extend the format slightly to allow recursion of files and
the specification of file and directory permissions when we recurse.

We include a basic shell-only version that can be used in all situations
for fallback installation, and a more robust Ruby version that's the
default if we have Ruby installed.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>