Add a set of mtree-based install rules
Add an mtree manifest for each set of existing files and generate a
complete manifest at the top level, which is ignored.  Once the manifest
is created, allow the `install-mtree` target to take the manifest and
perform the install.

Exclude the `.mtree` files from the installation using the older
approach.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>