spec/mtree: set up test files each iteration
We'd like to add some tests in the future about the results of modifying
some of the files in the test directory.  However, right now, this
directory is shared amongst all of the tests, which might result in
tests breaking if we did this.

To avoid this problem, let's set up the main code once and then set up
an independent set of test files within a temporary directory.  That
will allow us to modify the files as we wish, but is much faster since
we don't need to run `make install` multiple times.

The only major changes we need to make is that we need to chdir into the
new temporary directory and specify the path of the `dct-mtree` script
fully, since it will no longer be within our local directory.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>