playbooks/codespace-setup: avoid expensive chown when possible
In most codespaces, the ownership of files is already correct for most
of the files under /workspaces.  It's better to simply avoid the chown
if we can, which can be extremely expensive, if we believe the files are
already owned by the correct user.

However, when we do so, we must also make sure to do our chmod as root
as well, since Codespaces creates some files under /workspaces as root
and the operation will fail otherwise.

Signed-off-by: brian m. carlson <bk2204@github.com>