playbooks/codespace-setup: use Ansible native permission settings
Right now, we have to make two separate passes to set permissions, one
for chown, and one for chmod, which isn't very efficient.  Instead,
let's use the built-in Ansible functionality for setting the
permissions, which is much faster and is single-pass per directory.

In addition, make sure to set `follow` to false because we want to set
the ownership of symlinks, not traverse them.  If dotfiles are already
installed, following symlinks can result in an infinite loop in some
cases, causing the operation to fail.

Signed-off-by: brian m. carlson <bk2204@github.com>