playbooks/deploy-dotfiles: improve generality
Right now, this code is highly specific to a generic Debian or Ubuntu
instance.  However, it may be helpful to also run this in other
situations.

Allow specifying the checkouts directory and the branch, which could
allow this to be run in a work Codespace.  Additionally, make the call
to make work on FreeBSD as well, where GNU make is called `gmake`.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>