snippets: add a snippet about logical commits
In their role as a Git contributor, the present author has learned to
write nice commit messages with logical, bisectable commmits.  As a
result, it's possible for a pull request to contain a few nice
well-described cleanups and then the main change.

However, many projects are not expecting this, so it's likely that
they'll try to review the entire thing at once, which will likely be
more difficult to review.  In order to help them, we can provide a short
blurb of text that alerts them to the Git-style commits to ease their
review.

Do so, with a snippet under the name `logical-commits.`

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>