vimrc: set indent in Git commit messages to 4 spaces
When writing commit messages, it's often helpful to include code
snippets or quote some text.  On GitHub, this text ends up as Markdown,
but we also want it to be valid AsciiDoc when possible.

The easiest way to do this and include quoted text is to indent by four
spaces, since both Markdown and AsciiDoc consider that a verbatim block.
Therefore, regardless of the global indentation settings, let's set Git
commit messages to four spaces.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>