dct-mtree: improve compatibility with older versions of Ruby
The `:sort` keyword isn't available in the version of Ruby in Ubuntu
focal, and as such, this operation fails there.  Use a variable to store
the data and sort using the regular `sort` method on arrays.

Note that attempting to call `each` on the result of calling `sort` does
not appear to work and complains about the inability to compare String
and String.  As a result, we do not do that here.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>