Add `user.js` for Firefox
There are some Firefox settings, such as those having to do with
advertising, that we'd like to set on all systems.  Unfortunately,
Firefox doesn't use a fixed directory for its default profile, so we
have to find the right path by parsing the `profiles.ini` file and
extracting the default one. If such a profile exists, install the
`user.js` file.

Be sure to make the main mtree file and the `manifest.mtree` file depend
on the destination directory.  If we don't do that, then when running
RSpec from a tree from which we've already installed the dotfiles, we'll
end up with stale data and will try to install to a profile in the
temporary directory which doesn't exist, causing tests to fail.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>