spec: install from a clean source directory
When running rspec, the generated gitconfig file is generated.  This
means that our working tree contains these generated files, which causes
problems when running `make install` for real.  Because we don't want to
modify the user's configuration when running tests, be sure to copy the
file to a temporary directory, run `make clean` there, and only then
install the files.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>