Add a script to jump to certain locations
It's often useful to provide a way to jump to a certain URL or file
based on a text pattern.  For example, one might want to type `fr
bonjour` to look up the word “bonjour” in an English-French dictionary.

Add a script to allow doing so by reading from a Ruby configuration file
that matches certain patterns, and performs a given action based on
those.

Note that this differs from `dct-snip` because it has the ability to
perform regular expression matching and is not designed for full-fledged
scripting, but rather opening URLs or text files based on text patterns.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>