spec: fix failing zsh tests
The `cmd_with_exes` spec helper function is used for two purposes: one,
to create fake executable scripts to test path lookups, and two, to
set up a path for actual functional executable scripts that we want to
mock.  The first case requires that we create files which don't exist,
but the latter requires that we not truncate data.

Since none of the common Ruby IO modes actually specify exactly what we
want, let's just use the standard POSIX modes to indicate what we're
looking for.  This makes our tests work great in both cases.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>